Refactoring annotation processor: extract language model builder
